fix: address second round of PR review feedback
- Add test case for raw 16-byte binary input (errors, documenting that
  Scan expects UUID-formatted strings, not raw binary)
- Add test case for version-less UUID (accepted, confirming Scan is
  format-agnostic beyond structural validity)
- Assert roundtrip equality in ID wrapper tests
- Move ID wrapper tests to id_test.go
